+++ src/heat.js
@@ -1289,7 +1289,7 @@
     function importFromFilesSelected( bindingOptions ) {
         var input = createElementWithNoContainer( "input" );
         input.type = "file";
-        input.accept = ".json";
+        input.accept = ".json, .txt";
         input.multiple = "multiple";
 
         input.onchange = function() {
@@ -1316,7 +1316,7 @@
                     data[ storageDate ] += readingObject[ storageDate ];
                 }
             }
-
+            
             if ( filesCompleted.length === filesLength ) {
                 fireCustomTrigger( bindingOptions.onImport, bindingOptions.currentView.element );
                 renderControlContainer( bindingOptions );
@@ -1329,6 +1329,8 @@
 
             if ( fileExtension === _export_Type_Json ) {
                 importFromJson( file, onLoadEnd );
+            } else if ( fileExtension === _export_Type_Txt ) {
+                importFromTxt( file, onLoadEnd );
             }
         }
     }
@@ -1348,6 +1350,28 @@
 
             if ( jsonObject.parsed && isDefinedObject( jsonObject.result ) ) {
                 readingObject = jsonObject.result;
+            }
+        };
+    }
+
+    function importFromTxt( file, onLoadEnd ) {
+        var reader = new FileReader(),
+            readingObject = {};
+
+        reader.readAsText( file );
+
+        reader.onloadend = function() {
+            onLoadEnd( file.name, readingObject );
+        };
+    
+        reader.onload = function( e ) {
+            var lines = e.target.result.toString().split( _string.newLine ),
+                linesLength = lines.length;
+
+            for ( var lineIndex = 0; lineIndex < linesLength; lineIndex++ ) {
+                var line = lines[ lineIndex ].split( ":" );
+
+                readingObject[ line[ 0 ].trim() ] = parseInt( line[ 1 ].trim() );
             }
         };
     }